### 2018-11-27 #1100 @whedon
[PRE REVIEW]: sierra-local: A lightweight standalone application for drug resistance prediction
**Submitting author:** @ArtPoon (<a href="http://orcid.org/0000-0003-3779-154X">Art Poon</a>)
**Repository:** <a href="https://github.com/PoonLab/sierra-local" target ="_blank">https://github.com/PoonLab/sierra-local</a>
**Version:** v0.1.2
