### 2018-11-28 9ed9abf
Add line breaks to table to enforce hard wrap
