### 2018-11-28 f30e560
Fixing journal titles
