+++ paper/main.bib
@@ -3,7 +3,7 @@
 @article{shafer2006rationale,
   title={Rationale and uses of a public HIV drug-resistance database},
   author={Shafer, Robert W},
-  journal={The Journal of infectious diseases},
+  journal={J Infect Dis},
   volume={194},
   number={Supplement\_1},
   pages={S51--S58},
@@ -23,7 +23,7 @@
 @article{wilhelmson2018prevalence,
   title={Prevalence of HIV-1 pretreatment drug resistance among treatment na{\"\i}ve pregnant women in Bissau, Guinea Bissau},
   author={Wilhelmson, Sten and M{\aa}nsson, Fredrik and Lindman, Jacob Lopatko and Biai, Ansu and Esbj{\"o}rnsson, Joakim and Norrgren, Hans and Jansson, Marianne and Medstrand, Patrik and others},
-  journal={PloS one},
+  journal={PLoS ONE},
   volume={13},
   number={10},
   pages={e0206406},
@@ -34,7 +34,7 @@
 @article{arimide2018hiv,
   title={HIV-genetic diversity and drug resistance transmission clusters in Gondar, Northern Ethiopia, 2003-2013},
   author={Arimide, Dawit Assefa and Abebe, Almaz and Kebede, Yenew and Adugna, Fekadu and Tilahun, Tesfaye and Kassa, Desta and Assefa, Yibeltal and Balcha, Taye Tolera and Bj{\"o}rkman, Per and Medstrand, Patrik},
-  journal={PloS one},
+  journal={PLoS ONE},
   volume={13},
   number={10},
   pages={e0205446},
@@ -133,7 +133,7 @@ year = {2003}
 @article{tzou2017nucamino,
   title={NucAmino: a nucleotide to amino acid alignment optimized for virus gene sequences},
   author={Tzou, Philip L and Huang, Xiaoqiu and Shafer, Robert W},
-  journal={BMC Bioinformatics},
+  journal={BMC Bioinform},
   volume={18},
   number={1},
   pages={138},
@@ -162,7 +162,7 @@ year = {2003}
 @article{shafer2006,
 author = {Liu, T. F. and Shafer, R. W.},
 doi = {10.1086/503914},
-journal = {Clinical Infectious Diseases},
+journal = {Clin Infect Dis},
 keywords = {hiv-1},
 mendeley-groups = {sierralocal},
 month = {June},
@@ -197,7 +197,7 @@ year = {2003}
 author = {Bennett, Diane E. and Camacho, Ricardo J. and Otelea, Dan and Kuritzkes, Daniel R. and Fleury, Herv{\'{e}} and Kiuchi, Mark and Heneine, Walid and Kantor, Rami and Jordan, Michael R. and Schapiro, Jonathan M. and Vandamme, Anne Mieke and Sandstrom, Paul and Boucher, Charles A.B. and van de Vijver, David and Rhee, Soo Yon and Liu, Tommy F. and Pillay, Deenan and Shafer, Robert W.},
 doi = {10.1371/journal.pone.0004724},
 editor = {Nixon, Douglas F.},
-journal = {PLoS One},
+journal = {PLoS ONE},
 mendeley-groups = {sierralocal},
 month = {March},
 number = {3},
@@ -252,7 +252,7 @@ year = {2014}
   doi = {10.1371/journal.pone.0181357},
   month = {July},
   isbn = {1111111111},
-  journal = {PLoS One},
+  journal = {PLoS ONE},
   mendeley-groups = {sierralocal},
   number = {7},
   pages = {e0181357},
