### 2018-11-28 #1100 @ArtPoon
* The table exceeds page margins and some entries are running together.  I've added some `<br/>` tags in the table to try to improve this formatting.
* Journal titles are inconsistently abbreviated, and there was no automatic adjustment of capitalization.  I have manually edited the title entries to make this consistent throughout the reference list.
