### 2018-11-29 d0263dc
Dropping the table
JOSS table formatting doesn't seem to accept manual line breaks with `<br/>`.
I don't think this information is really necessary anyway.
Added link to bioRxiv preprint.
