+++ paper/main.bib
@@ -1,5 +1,14 @@
 % Encoding: UTF-8
 
+@article{ho2018sierra,
+  title={sierra-local: A lightweight standalone application for secure HIV-1 drug resistance prediction},
+  author={Ho, Jasper C and Ng, Garway T and Renaud, Mathias and Poon, Art FY},
+  journal={bioRxiv},
+  pages={393207},
+  year={2018},
+  publisher={Cold Spring Harbor Laboratory}
+}
+
 @article{shafer2006rationale,
   title={Rationale and uses of a public HIV drug-resistance database},
   author={Shafer, Robert W},
