### 2018-12-19 #1100 @george-githinji
Thank you @azneto @betteridiot.
 @pjotrp would you mind taking this on as an editor?
