### 2018-12-28 #1100 @pjotrp
Thanks @ArtPoon, this is an interesing submission. @azneto do you mind being the primary reviewer? No rush.
