### 2019-01-18 #1186 @whedon
[REVIEW]: sierra-local: A lightweight standalone application for drug resistance prediction
**Submitting author:** @ArtPoon (<a href="http://orcid.org/0000-0003-3779-154X">Art Poon</a>)
**Repository:** <a href="https://github.com/PoonLab/sierra-local" target ="_blank">https://github.com/PoonLab/sierra-local</a>
**Version:** v0.1.2

@azneto , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
