+++ paper/main.bib
@@ -1,7 +1,7 @@
 % Encoding: UTF-8
 
 @article{ho2018sierra,
-  title={sierra-local: A lightweight standalone application for secure HIV-1 drug resistance prediction},
+  title={sierra-local: A lightweight standalone application for secure {HIV}-1 drug resistance prediction},
   author={Ho, Jasper C and Ng, Garway T and Renaud, Mathias and Poon, Art FY},
   journal={bioRxiv},
   pages={393207},
@@ -10,7 +10,7 @@
 }
 
 @article{shafer2006rationale,
-  title={Rationale and uses of a public HIV drug-resistance database},
+  title={Rationale and uses of a public {HIV} drug-resistance database},
   author={Shafer, Robert W},
   journal={J Infect Dis},
   volume={194},
@@ -21,7 +21,7 @@
 }
 
 @article{rasmussen2017external,
-  title={External introductions helped drive and sustain the high incidence of HIV-1 in rural KwaZulu-Natal, South Africa},
+  title={{External introductions helped drive and sustain the high incidence of HIV-1 in rural KwaZulu-Natal, South Africa}},
   author={Rasmussen, David and Wilkinson, Eduan and Vandormael, Alain and Tanser, Frank and Pillay, Deenan and Stadler, Tanja and de Oliveira, Tulio},
   journal={bioRxiv},
   pages={119826},
@@ -30,7 +30,7 @@
 }
 
 @article{wilhelmson2018prevalence,
-  title={Prevalence of HIV-1 pretreatment drug resistance among treatment na{\"\i}ve pregnant women in Bissau, Guinea Bissau},
+  title={{Prevalence of HIV-1 pretreatment drug resistance among treatment na{\"\i}ve pregnant women in Bissau, Guinea Bissau}},
   author={Wilhelmson, Sten and M{\aa}nsson, Fredrik and Lindman, Jacob Lopatko and Biai, Ansu and Esbj{\"o}rnsson, Joakim and Norrgren, Hans and Jansson, Marianne and Medstrand, Patrik and others},
   journal={PLoS ONE},
   volume={13},
@@ -41,7 +41,7 @@
 }
 
 @article{arimide2018hiv,
-  title={HIV-genetic diversity and drug resistance transmission clusters in Gondar, Northern Ethiopia, 2003-2013},
+  title={{HIV-genetic diversity and drug resistance transmission clusters in Gondar, Northern Ethiopia, 2003-2013}},
   author={Arimide, Dawit Assefa and Abebe, Almaz and Kebede, Yenew and Adugna, Fekadu and Tilahun, Tesfaye and Kassa, Desta and Assefa, Yibeltal and Balcha, Taye Tolera and Bj{\"o}rkman, Per and Medstrand, Patrik},
   journal={PLoS ONE},
   volume={13},
@@ -52,7 +52,7 @@
 }
 
 @article{solis2018open,
-  title={An open-source k-mer based machine learning tool for fast and accurate subtyping of HIV-1 genomes},
+  title={An open-source k-mer based machine learning tool for fast and accurate subtyping of {HIV}-1 genomes},
   author={Solis-Reyes, Stephen and Avino, Mariano and Poon, Art and Kari, Lila},
   journal={bioRxiv},
   pages={362780},
@@ -80,7 +80,7 @@ number = {2},
 pages = {98--101},
 pmid = {22286876},
 publisher = {Karger Publishers},
-title = {{The HIVdb System for HIV-1 Genotypic Resistance Interpretation}},
+title = {The {HIVdb} System for {HIV}-1 Genotypic Resistance Interpretation},
 url = {http://www.ncbi.nlm.nih.gov/pubmed/22286876},
 volume = {55},
 year = {2012}
@@ -113,7 +113,7 @@ number = {4},
[…]
