+++ paper/main.bib
@@ -360,7 +360,7 @@ year = {2002}
 @InProceedings{patil2014big,
   author       = {Patil, Harsh Kupwade and Seshadri, Ravi},
   title        = {Big data security and privacy issues in healthcare},
-  booktitle    = {Big Data (BigData Congress), 2014 IEEE International Congress on},
+  booktitle    = {{Big Data (BigData Congress), 2014 IEEE International Congress on}},
   year         = {2014},
   pages        = {762--765},
   organization = {IEEE},
