### 2019-01-18 #1186 @ArtPoon
Thanks for your help @pjotrp.  I confirm that I am unable to tick the checkboxes above.

I've read through the current article proof.  The only issue I've noticed is that the capitalization of some article and journal titles was not properly protected in the `.bib` file.  I have made corrections - is it ok for me to re-render the proof?
