### 2019-01-19 #1186 @pjotrp
Of course. '@whedon generate pdf' will do it.
