### 2019-01-22 #1186 @ArtPoon
OK @pjotrp, I've checked the revised PDF output and it looks OK to me.
