### 2019-01-22 #1186 @pjotrp
Thanks @ArtPoon. @azneto you can start review here.
