### 2019-01-22 PoonLab/sierra-local#52 @azneto
[JOSS review] community guidelines
From the JOSS review criteria: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support?". Please add these items to your documentation. Consider using a Issue template and adding a CONTRIBUTING.md file.
