### 2019-01-22 PoonLab/sierra-local#53 @azneto
[JOSS review] sierrapy reference
The paper.md reports the comparison of sierra-local and sierrapy, but there's no precise reference to the sierrapy repo/url.
