### 2019-01-22 #1186 @azneto
@ArtPoon, congratulations! This is a very useful tool to run the analyses on HIVdb. It considerably improves the user experience in every way by allowing to run analyses locally. I've ran the software a few times during the last days using Ubuntu. The software is well documented and runs really fast.

Though the documentation for the end users is really good and comprehensive, the code itself is not so inviting. It doesn't follow the PEP8 (Style Guide for Python Code) and the PEP20 (The Zen of Python). There are variables and libraries that were created/loaded and never used (eg. definitions, database, comments, pathlib.Path, requests, default_grange). The software flake8 can find those and also check if the code is compatible with the PEP8, in future releases. I also recommend using typing (https://docs.python.org/3/library/typing.html).

The software should be ACCEPTED with minor revisions:
https://github.com/PoonLab/sierra-local/issues/52
https://github.com/PoonLab/sierra-local/issues/53
