### 2019-01-22 #1186 @pjotrp
Also a good idea to add a few lines on community guidelines. How can people contribute. Sometimes it happens ;)
