### 2019-01-22 #1186 @azneto
@ArtPoon, the style is important but not required for this review. Just focus on the two pending issues and let the other improvements for the next releases.
