### 2019-01-22 #1186 @ArtPoon
Thanks @azneto for your review!  As you suggest, I will address the issues you've opened for the paper, and then in the longer term work on improving the code style.  I'll also add a `CONTRIBUTING` file as suggested and report back here.
