### 2019-01-22 b6f525d
Create CONTRIBUTING.md
