+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+## How to contribute
+
+If you would like to contribute to the development and maintenance of *sierra-local*, great!  We are trying to adhere to the following style conventions for Python code:
+
+* Use a block quote at the top of a new script to briefly describe its purpose.
+* Indent with 4 spaces.
+* Variable names should use lowercase and underscores.
+* Function names use camelCase.
+* Functions should have a block quote in the first few lines to provide a description and to list input parameters and return values.
+
+## How to report issues
+
+* We use the GitHub issue tracker to report and coordinate work on issues.  
+* If you encounter a bug, please search the issue tracker for related issues before creating a new issue.
+* Please provide the following information:
+  * the *sierra-local* release version number or commit
+  * operating system
+  * Python version
+  * a brief description of the problem
+  * if appropriate, an informative excerpt of traceback or error messages returned by the program, enclosed in triple ticks.
