### 2019-01-22 20f2d7d
Update CONTRIBUTING.md
