### 2019-01-22 f726c92
Update CONTRIBUTING.md
