### 2019-01-22 PoonLab/sierra-local#52 @ArtPoon
Started writing `CONTRIBUTING.md` file.
