### 2019-01-22 d35706e
Update paper.md
Added URL to SierraPy repo on GitHub.
Fixes #53
