### 2019-01-23 335ac5d
Update CONTRIBUTING.md
