+++ CONTRIBUTING.md
@@ -6,7 +6,7 @@ If you would like to contribute to the development and maintenance of *sierra-lo
 * Indent with 4 spaces.
 * Variable names should use lowercase and underscores.
 * Constants are named with uppercase and underscores, *e.g.*, `SEQUENCE_SHRINKAGE_WINDOW`
-* Function names use camelCase.
+* Function names should use camelCase.
 * Functions should have a block quote in the first few lines to provide a description and to list input parameters and return values.
 
 ## How to report issues
@@ -20,3 +20,7 @@ If you would like to contribute to the development and maintenance of *sierra-lo
   * a brief description of the problem
   * if appropriate, an informative excerpt of traceback or error messages returned by the program, enclosed in triple ticks.
 * If you are requesting a new feature, please create a new issue and mark it with an *enhancement* tag.
+
+## How to get help
+* If you have a [GitHub account](https://github.com/join), you are welcome to post an issue on the [sierra-local issue tracker](https://github.com/PoonLab/sierra-local/issues).  
+* If you are an end-user at an academic organization or clinical laboratory, you can contact [Art](https://www.schulich.uwo.ca/pathol/people/bios/faculty/poon_art.html) (currently package maintainer) at his work e-mail (see linked faculty webpage).
