### 2019-01-23 df9ea9d
Update CONTRIBUTING.md
