### 2019-01-23 1b4f195
Update CONTRIBUTING.md
