### 2019-01-23 e2680de
Update CONTRIBUTING.md
