### 2019-01-23 #1186 @ArtPoon
Hi @azneto and @pjotrp, I've addressed and closed the issues.
