### 2019-01-24 c00b2a8
Update main.bib
Added missing DOIs, replaced Rasmussen preprint citation with published article.
