+++ paper/main.bib
@@ -6,7 +6,8 @@
   journal={bioRxiv},
   pages={393207},
   year={2018},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/393207}
 }
 
 @article{shafer2006rationale,
@@ -17,16 +18,20 @@
   number={Supplement\_1},
   pages={S51--S58},
   year={2006},
-  publisher={University of Chicago Press}
+  publisher={University of Chicago Press} ,
+  doi={10.1086/505356}
 }
 
 @article{rasmussen2017external,
-  title={{External introductions helped drive and sustain the high incidence of HIV-1 in rural KwaZulu-Natal, South Africa}},
-  author={Rasmussen, David and Wilkinson, Eduan and Vandormael, Alain and Tanser, Frank and Pillay, Deenan and Stadler, Tanja and de Oliveira, Tulio},
-  journal={bioRxiv},
-  pages={119826},
-  year={2017},
-  publisher={Cold Spring Harbor Laboratory}
+  title={Tracking external introductions of HIV using phylodynamics reveals a major source of infections in rural KwaZulu-Natal, South Africa},
+  author={Rasmussen, David A and Wilkinson, Eduan and Vandormael, Alain and Tanser, Frank and Pillay, Deenan and Stadler, Tanja and de Oliveira, Tulio},
+  journal={Virus Evolution},
+  volume={4},
+  number={2},
+  pages={vey037},
+  year={2018},
+  publisher={Oxford University Press},
+  doi={10.1093/ve/vey037}
 }
 
 @article{wilhelmson2018prevalence,
@@ -37,7 +42,8 @@
   number={10},
   pages={e0206406},
   year={2018},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0206406}
 }
 
 @article{arimide2018hiv,
@@ -48,7 +54,8 @@
   number={10},
   pages={e0205446},
   year={2018},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0205446}
 }
 
 @article{solis2018open,
@@ -341,7 +348,6 @@ number = {2},
 pages = {123--129},
 pmid = {12212924},
 title = {{A genotypic drug resistance interpretation algorithm that significantly predicts therapy response in HIV-1-infected patients}},
-url = {http://www.sante.gouv.fr/htm/publication/index.htm;},
 volume = {7},
 year = {2002}
 }
@@ -364,6 +370,7 @@ year = {2002}
   year         = {2014},
   pages        = {762--765},
   organization = {IEEE},
+  doi = {10.1109/BigData.Congress.2014.112}
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
