### 2019-01-24 #1186 @ArtPoon
@pjotrp, I have deposited the current release of the repository to Zenodo, obtained and posted the DOI above.  I also found and restored some missing DOIs for cited references in the paper.  Unfortunately a couple of the journals we referenced do not provide DOIs (namely *AIDS* and *Antiviral Therapy*).

Further, I replaced a citation to a bioRxiv preprint with the recently published version of that manuscript (Rasmussen et al. 2018), and I removed a broken URL associated with Van Laethem et al. (2002).
