### 2019-01-25 #1186 @pjotrp
Hi @ArtPoon, can you post the zenodo or figshare doi of the software release here?
