### 2019-01-25 #1186 @ArtPoon
I did post the Zenodo DOI in this issue.
